Add confirmation prompt for deleting posts/comments

+++ admin-frontend/src/Components/BlogList.js
@@ -3,11 +3,17 @@ import {Link} from 'react-router-dom'
 
 function BlogListItem(props) {
     const deleteBlog = () => {
-        fetch(`/api/posts/${props.id}`, {
-            method: 'DELETE'
-        }).then(res => res.json()).then(data => { 
-            console.log(data)
-        })
+        const ok = window.confirm('Are you sure you want to delete this post and all associated comments?')
+        if(ok) {
+            fetch(`/api/posts/${props.id}`, {
+                method: 'DELETE'
+            }).then(res => res.json()).then(data => { 
+                console.log(data)
+            })
+        } else {
+            return
+        }
+        
     }
 
     return(
