+++ admin-frontend/src/Components/Comment.js
@@ -2,9 +2,15 @@ import React from 'react';
 
 function Comment(props) {
     const deleteComment = () => {
-        fetch(`/api/posts/${props.post._id}/comments/${props.id}`, {
-            method: 'DELETE',
-        }).then(res => res.json()).then(data => console.log(data))
+        const ok = window.confirm('Are you sure you want to delete this comment?')
+        if(ok) {
+            fetch(`/api/posts/${props.post._id}/comments/${props.id}`, {
+                method: 'DELETE',
+            }).then(res => res.json()).then(data => console.log(data))
+        } else {
+            return 
+        }
+        
         
     }
